Restructure components and update project assets
Moved and reorganized component files into component-based directories for improved maintainability.
