+++ src/components/filter/SportFilter.tsx
@@ -1,13 +1,17 @@
 import * as React from 'react';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from './ui';
-import {SportFilterProps} from "./models";
-import {SPORTS_PLACEHOLDER} from "../app.constants";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../../ui';
 
-export function SportFilter({ selectedSport, onSportChange, sports }: SportFilterProps) {
+export interface SportFilterProps {
+    selectedSport: string;
+    onSportChange: (sport: string) => void;
+    sports: string[];
+}
+
+export default function SportFilter({ selectedSport, onSportChange, sports }: SportFilterProps) {
   return (
     <Select value={selectedSport} onValueChange={onSportChange}>
       <SelectTrigger>
-        <SelectValue placeholder={SPORTS_PLACEHOLDER} />
+        <SelectValue placeholder="Filter by sport" />
       </SelectTrigger>
       <SelectContent>
         <SelectItem value="all">All Sports</SelectItem>
