+++ src/components/league/LeagueCard.tsx
@@ -1,8 +1,13 @@
 import * as React from 'react';
-import {Card, CardContent, CardHeader, CardTitle, Badge} from './ui';
-import {LeagueCardProps} from "./models";
+import {Card, CardContent, CardHeader, CardTitle, Badge} from '../../ui';
+import {League} from "../../types";
 
-export function LeagueCard({league, onClick}: LeagueCardProps) {
+export interface LeagueCardProps {
+    league: League;
+    onClick: () => void;
+}
+
+export default function LeagueCard({league, onClick}: LeagueCardProps) {
     return (
         <Card
             className="cursor-pointer hover:shadow-lg transition-shadow duration-200 hover:bg-accent/50"
