--- src/components/models/index.ts
@@ -1,4 +1,0 @@
-﻿export * from './league';
-export * from './search';
-export * from './season';
-export * from './sport';
