--- src/components/models/search.ts
@@ -1,4 +1,0 @@
-﻿export interface SearchBarProps {
-    searchTerm: string;
-    onSearchChange: (term: string) => void;
-}
