--- src/components/models/season.ts
@@ -1,8 +1,0 @@
-﻿import {League, SeasonBadge} from "../../types";
-
-export interface SeasonBadgeDialogProps {
-    league: League | null;
-    seasonBadge: SeasonBadge | null;
-    loading: boolean;
-    onClose: () => void;
-}
