--- src/components/models/sport.ts
@@ -1,5 +1,0 @@
-﻿export interface SportFilterProps {
-    selectedSport: string;
-    onSportChange: (sport: string) => void;
-    sports: string[];
-}
