+++ src/components/search/SearchBar.tsx
@@ -1,14 +1,20 @@
 import * as React from 'react';
 import { Search } from 'lucide-react';
-import { Input } from './ui';
-import { SearchBarProps } from './models';
-export function SearchBar({ searchTerm, onSearchChange }: SearchBarProps) {
+import { Input } from '../../ui';
+
+export interface SearchBarProps {
+    searchTerm: string;
+    onSearchChange: (term: string) => void;
+    placeholder: string;
+}
+
+export default function SearchBar({ searchTerm, onSearchChange, placeholder = "" }: SearchBarProps) {
   return (
     <div className="relative">
       <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-4 w-4" />
       <Input
         type="text"
-        placeholder="Search leagues by name..."
+        placeholder={placeholder}
         value={searchTerm}
         onChange={(e) => onSearchChange(e.target.value)}
         className="pl-10"
