+++ src/components/search/search.ts
@@ -1,0 +1,1 @@
+﻿
